:pushpin: add sdkmanrc file to set sdk versions (jdk and maven) in a machine-readable way
Use Amazon Corretto as a JDK for now because they actually have a Java 21 release available.

This is supported by IntelliJ and, of course, by sdkman. You can also configure sdkman to automatically switch the sdk(s) when entering the project directory in your shell. See the comment in the file.