🐛 fix bug happening when using virtual threads
When running the old code with -Djdk.tracePinnedThreads=full, we can see that virtual threads are pinned. That is being caused by the ConcurrentHashMap's "computeIfAbsent" method containing a synchronized block.

We can get away with an "un-thread-safe" implementation here, because we rely on explicit locking of the memory anyway.

The key takeaway here is: Do not do I/O in the "computeIfAbsent" method of a ConcurrentHashMap …